--- Project/Assets/Prefabs/Collectables/Charge.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4c4f9b0a3dbd8f146b75730c83059199
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
