+++ Project/Assets/Prefabs/Collectables/Prism.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d13a8331121672a48bd303bdbbbd6241
+guid: 4c4f9b0a3dbd8f146b75730c83059199
 PrefabImporter:
   externalObjects: {}
   userData:
