+++ Project/Assets/Prefabs/Collectables/PrismDrop.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d13a8331121672a48bd303bdbbbd6241
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
