--- Project/Assets/Scripts/Collectables/Charge.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 40608e8d79cef0c4f8ad43559f96736a
