+++ Project/Assets/Scripts/Collectables/Prism.cs.meta
@@ -1,11 +1,2 @@
 fileFormatVersion: 2
-guid: 7c74cc291d93f2f47852072cdc31faa2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+guid: 40608e8d79cef0c4f8ad43559f96736a
