+++ Project/Assets/Scripts/Collectables/PrismDrop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c74cc291d93f2f47852072cdc31faa2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
