+++ Project/Assets/Sprites/Test/BetaAssets/Beta_Level_1.spriteatlas
@@ -27,8 +27,8 @@ SpriteAtlas:
       blockOffset: 1
       allowAlphaSplitting: 0
       enableRotation: 1
-      enableTightPacking: 0
-      enableAlphaDilation: 1
+      enableTightPacking: 1
+      enableAlphaDilation: 0
     secondaryTextureSettings: {}
     variantMultiplier: 1
     packables:
